+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 
+- Add ExtendedFields model [#1503](https://github.com/open-apparel-registry/open-apparel-registry/pull/1503)
+
 ### Changed
 
 ### Deprecated